fix: make lambda function depend on the Cloudwatch log group
If the lambda function writes logs when the log group doesn't exist,
then AWS will create it, and Terraform will error with "log group
already exists" on the following `terraform apply`.

It's possible to trigger this case when running `terraform destroy`
when the lambda function is being invoked: the log group will be
deleted by Terraform and recreated by AWS before the lambda function
is destroyed.

This change imposes an ordering between the lambda function and the
log group creation/destruction, avoiding this race condition.