Rebuild deployment package on architecture change
Previously the CPU architecture configured for an AWS Lambda Function
wasn't taken into account when generating the file name hash for the
deployment package. This meant that changing the CPU architecture of an
AWS Lambda Function did not trigger a rebuild of the deployment package.
As deployment packages might contain CPU-specific binaries, this
could've caused deployment breaking AWS Lambda Functions. This commit
fixes that by taking the CPU architecture into account for the file name
hash of the deployment package as well.

One minor annoyance of the current implementation is that switching from
an unconfigured CPU architecture to `x86-64` will also result in a
rebuild of the deployment package, as that results in a switch of the
value for the architecture from `null` to `["x86-64"]`. However, I'm not
sure what'd be the best way to fix that and as it only happens when
intentionally changing the architecture attribute, I believe it's fine
as is as well.