fix(lambda-permissions): create before destroy
When any parameters contain dynamic elements, e.g. `data.aws_region.name`, and
that is evaluated in an intermediate module, it is not known until apply time,
which causes the permission to he removed and added.  During this window,
Cloudwatch Logs will see the errors and stop sending events for 10 minutes.

By switching to a statement_id_prefix and using a create_before_destroy
lifecycle we avoid such interruptions.