Avoids short-circuit issue where fileexists returns inconsistent results
In old versions of terraform, the ternary operator did not short-circuit,
and so the `was_missing` logic was a hack to workaround both paths executing
on every execution. The module needed the hack to avoid failing When the
package existed previously vs when it did not and was being created by the
module itself.

In recent-ish version of terraform, including the module min version 1.5.7,
the ternary operator will shortcircuit, so it will only execute the true OR
false logic, rather than both. This makes the `was_missing` logic unnecessary.

Further, in most recent versions of terraform, at least since 1.13.0, the
`was_missing` logic resulted in a failure when separately building the package
and then attempting to use the package in the same workflow. The failure occurred
because terraform became more strict at enforcing that the return values of the
`fileexists()` function be consistent between plan and apply phases.

Fixes https://github.com/terraform-aws-modules/terraform-aws-lambda/issues/698